Generic connector added to reduce code redundancy
